requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-hosts==4.0
django-nested-admin==3.2.4
django-rest-swagger==2.2.0
django-tinymce==2.8.0
djangorestframework==3.11.0
entrypoints==0.3
Faker==0.9.1
flake8==3.7.9
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.2.0
openapi-codec==1.3.2
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-flake8==1.0.4
python-dateutil==2.8.1
python-monkey-business==1.0.0
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
